requirement.txt
py-cord==2.4.1
python-dotenv==1.0.1
google-generativeai==0.3.2
